requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests>=2.26.0,<3.0.0
aiofiles>=0.7.0,<0.8.0
numpy>=1.21.2,<2.0.0
imageio>=2.9.0,<3.0.0
opencv-python
kornia==0.6.12
tqdm>=4.62.3,<5.0.0
torch-scatter>=2.0.8
Pillow>=8.3.2
einops>=0.3.0
commentjson>=0.9.0
pandas>=1.3.3
tensorboard>=2.6.0
